refactor: move auth dummy schema into separate file
